+++ app/src/main/java/com/nafanya/mp3world/data/media_store/MediaStoreReader.kt
@@ -1,6 +1,5 @@
 package com.nafanya.mp3world.data.media_store
 
-import androidx.annotation.WorkerThread
 import com.nafanya.mp3world.core.wrappers.song.local.LocalSong
 
 interface MediaStoreReader {
